move collections related code to a specific folder
